Use BigDecimal for money

+++ src/main/kotlin/dr/kotliners/kotlinbackend/model/Account.kt
@@ -1,5 +1,6 @@
 package dr.kotliners.kotlinbackend.model
 
+import java.math.BigDecimal
 import java.util.*
 import kotlin.collections.HashSet
 
@@ -7,19 +8,19 @@ data class Account(
     val id: Long,
     val userId: Int,
     val currency: Currency,
-    var amount: Double,
+    var amount: BigDecimal,
     val transactions: HashSet<Transaction>
 )
 
 data class Transaction(
     val id: Long,
     val accountId: Long,
-    val value: Double,
+    val value: BigDecimal,
     val type: TransactionType,
     val date: Long
 ) {
     companion object {
-        fun depositTransaction(accountId: Long, deposit: Double): Transaction {
+        fun depositTransaction(accountId: Long, deposit: BigDecimal): Transaction {
             val transactionId = UUID.randomUUID().leastSignificantBits
             return Transaction(
                 id = transactionId,
@@ -30,7 +31,7 @@ data class Transaction(
             )
         }
 
-        fun transferTransaction(accountId: Long, value: Double): Transaction {
+        fun transferTransaction(accountId: Long, value: BigDecimal): Transaction {
             val transactionId = UUID.randomUUID().mostSignificantBits
             return Transaction(
                 id = transactionId,
